requirements.txt
# Web Framework
Flask==3.0.0

# MongoDB Database
pymongo==4.6.1

# Payment Gateway
razorpay==1.4.1

# Document Processing
python-docx==1.1.0
PyMuPDF==1.23.8

# Image Processing
Pillow==10.1.0

# Google AI APIs
google-generativeai==0.8.3
requests==2.31.0

# Security
werkzeug==3.0.1

# Environment Variables (optional but recommended)
python-dotenv==1.0.0
